requirements.txt
asgiref==3.4.1
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.45.0
django-pandas==0.6.4
gunicorn==20.1.0
idna==3.2
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
whitenoise==5.3.0
